requirements.txt
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.3
jwt==1.3.1
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
SQLAlchemy==2.0.29
Werkzeug==3.0.2
